CHROMIUM: ARM: dts: Fix WP on mighty
Mighty has a full sized SD card with the WP line going to the AP just fine.
There's no reason it should be disabled. This reorganized the code a little,
enables it on just mighty (proper, not the .dtsi).

Clarification on how the states work:
SD card on RW position = Connector switch asserted = switch pulls the signal to low
SD card on RO position = Connector switch not asserted = pullup holds it high

Also moved sdmmc_wp_gpio since it's only applicable on mighty.
No reason for the other boards to initialize this pin.

BUG=chrome-os-partner:40153
TEST=Set SD card to write protect, dd something over, it should not error saying it's RO
TEST=Non write protect should still work

Change-Id: I859193829f2b1d1e9cb01d308b025e6e4009250c
Signed-off-by: Alexandru M Stan <amstan@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/271629
Reviewed-by: Douglas Anderson <dianders@chromium.org>