CHROMIUM: mmc: dw_mmc: Only send stop commands for abort during tuning
It's not expected that we'll send a stop command after a tuning
command.  Doing so will always result in an error, so there's no good
reason to do it when a tuning command worked with no errors.

We still _do_ need to make sure to send a stop/abort command when we
get an error, though.  We'll follow the way that upstream did that in
(6c2c650 mmc: dw_mmc: Convert to mmc_send_tuning()) by modifying
dw_mci_prep_stop_abort().

This patch should be reverted if/when we move to mmc_send_tuning()
since then we can just pick the upstream pick and get this.

BUG=chrome-os-partner:40498
TEST=Reboot tests; suspend_resume tests; mmc_test

Change-Id: I7d9ab26d663c5c7a74e4747b3e8828449002fec1
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/272207
Reviewed-by: Alexandru Stan <amstan@chromium.org>
Reviewed-by: Andrew Bresticker <abrestic@chromium.org>