FROMLIST: ARM: rockchip: restore dapswjdp after suspend
In the commit (0ea001d ARM: rockchip: disable dapswjdp during suspend)
we made the assumption that we didn't need to restore dapswjdp after
suspend because "the MASKROM will enable it back".

It turns out that's not a safe assumption.  In some cases (pending
interrupts) it's possible that the WFI might act as a no-op and the
MaskROM will never run.  Since we're changing the bit, we should
restore it ourselves.

BUG=chrome-os-partner:40447
TEST=Compare 'mem r 0xff740040' before/after suspend w/ no cpu_do_idle()

(am from https://patchwork.kernel.org/patch/6448961/)
Change-Id: If7e804aa9c650e5d973e7500569cdbeaeb488c99
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/272443
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>