mfd/cros_ec: add debugfs, console log file
If the EC supports the new CONSOLE_READ command type, then we
place a console_log file in debugfs for that EC device which allows
us to grab EC logs. The kernel will poll every 30 seconds for the
log and keep its own buffer, but userspace should grab this and
write it out to some logs which actually get rotated.

BUG=chromium:492721
TEST=on samus, cat /sys/kernel/debug/cros_ec/console_log and see
  that it has data, cat again and see that it blocks until the
  kernel has polled the EC and data is available

Change-Id: I3051a44fa530fa6ea723632f0378942734cb0685
Signed-off-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/276693
Reviewed-by: Kevin Cernekee <cernekee@chromium.org>
Reviewed-by: Gwendal Grignou <gwendal@chromium.org>