CHROMIUM: ARM: dts: danger: Use sdmmc_cd pullup to workaround a hardware problem
sdmmc_cd is wrongly pulled up externally to vccio_sd, which gets turned off
when there's no card. This causes the sdmmc driver to keep loop thinking
there's a card present (then turning the voltages off when it realized there
isn't a card).

We can enable the internal pullup (which gets pulled up from the proper rail)
and rely on the fact that the internal pullup is smaller than the external,
so it will win.

BUG=chrome-os-partner:40698
TEST=No more "rockchip-iodomain io-domains.18: Setting to 3300000 done" spam

Change-Id: I9f06734616667046365689de9da6d07f360b63a3
Signed-off-by: Alexandru M Stan <amstan@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/273869
Reviewed-by: Douglas Anderson <dianders@chromium.org>