CHROMIUM: drm: bridge/dw_hdmi: adjust some mode in fixup funcs
Because we can not find an NPLL rate which have generate accurate
136.75MHz and good jitter performance both, so after disscuss with doug
and stephane, we plan to range dclk_vop0 to 136.80MHz, so dclk_vop0 could
good jitter performance(118.790ps). In order the make clk driver could
chose the 1368MHz when we set dclk_vop0 to 136.75MHz, we have to force
dclk_vop0 to 136.80MHz (same reason to 146.25MHz, which would otherwise
pick 585MHz and bad jitter).

BUG=chrome-os-partner:39069
TEST=jerry board, hack adjust mode clock in fixup function, then we
     can output those hack value from bridge->mode_set function.

Change-Id: Ie3714db51522d397fbb7442512540cfbb724327a
Signed-off-by: Yakir Yang <ykk@rock-chips.com>
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/273648
Reviewed-by: Dominik Behr <dbehr@chromium.org>