CHROMIUM: drm: rockchip/dw_hdmi-rockchip: expand the informal mpll config
There are serval pixel clock which we can not find out
in the Synopsys/Designerware HDMI databook, so we must
build those params by ourselves.

As we already have made an good jitter performance from
the dclk_vop0, so our IC guys suggest that we only need
to find a couple of params which tmds equal to pixel
clock and close the clock rate we want. For example, if
we want build mpll param for 28.32MHz, then we can copy
the exist 27.00MHz with color depth equal to 8bit.

Here are the new expand tmds list:
28.32MHz, 30.24MHz, 57.28MHz, 172.80MHz

BUG=chrome-os-partner:39069
TEST=jerry board, connect to HDMI analytical instruments,
     make jerry output overhead pixel clock, and then
     measure the clock jitter.

Change-Id: If490b2b2db684481f992b11c4c123ff4af56e4ad
Signed-off-by: Yakir Yang <ykk@rock-chips.com>
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/273993