CHROMIUM: cpufreq: interactive: only apply interactive boost when enabled
Only apply the interactive boost when the interactive governor is
enabled.  This seems like the right thing to do.

It also closes a race at boot between when the input handler is registered
(in CPUFREQ_GOV_POLICY_INIT), and when we set non-NULL per-cpu policies
(in CPUFREQ_GOV_START).  If an input event arrives during this window,
cpufreq_interactive_boost() will OOPS when trying to access a NULL
pcpu->policy.

Signed-off-by: Daniel Kurtz <djkurtz@chromium.org>

BUG=chrome-os-partner:40664
TEST=boot oak w/ USB plugged and try to boot from USB by
 repeatedly pressing (or holding) Ctrl-U at the firmware screen...
 => system boots w/out OOPS

Prev-Change-Id: I5f4a7320683eada099f9a4253e3d6b0f03057fe8
Prev-Reviewed-on: https://chromium-review.googlesource.com/273501
Change-Id: Idfae87b9e867ed1bb99e3f78007ca570a096ac52
Reviewed-on: https://chromium-review.googlesource.com/273984
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Tested-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>
Commit-Queue: Dmitry Torokhov <dtor@chromium.org>