CHROMIUM: Input: atmel_mxt_ts - implement power up procedure for S2R
On some ARM systems kernel should control power rails to which the
touchpad or touchscreen is attached. Previously we only executed proper
power-up sequence at boot time and expected to regulators to stay
powered on suring S2R. This change adds proper power down/power up
sequence to the suspend and resume handlers in the driver and allows us
to turn regulators off and on during system sleep transitions.

Additionally we ensure that inhibit/uninhibit and open/close to not
clash with each other with regard to controller power state and that we
keep the device in low powered mode while it is closed (or inhibited).
Note that even device is closed or inhibited we still keep power rails
on to allow sysfs access (for example for config or firmware flashing).

BUG=chrome-os-partner:38094,chrome-os-partner:38175
TEST=powerd_dbus_suspend, then wakeup system, verify devices work
normally; execute similar test with closing/opening the lid; also
exercise "inhibit" interface. Execute tests on devices that have either
Atmel touchscreen or touchpad or both.

Change-Id: I1663eb486089b80835bcfecc717ee04e1dead69a
Signed-off-by: Dmitry Torokhov <dtor@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/275668
Reviewed-by: Douglas Anderson <dianders@chromium.org>