CHROMIUM: cros_usbpd-charger: report dual role devices as type USB
The EC can report whether a device is dual role or not. If the connected
device can be a power source and is dual role capable, report the power
supply type as USB, otherwise report the type as Mains.

Also, for the transitionary phase when the EC is negotiating power
capabilites with the newly connected device since the power capabilities
are unknown, report the power capability as 20V @ 5A.

BUG=chromium:459412
TEST=Connect various kinds of devices to the Type-C port and then under
/sys/class/power_supply/CROS_USB_PD_CHARGER0 for left port or
/sys/class/power_supply/CROS_USB_PD_CHARGER1 for right port:
1. for a dedicated charger such as a Zinger verify that:
- online is 1
- status is Charging
- type is Mains

2. for DP/HDMI Type-C dongles verify that:
- online is 0
- status is Not Charging
- type is Unknown

3. for dual-role Type-C devices such as Samus verify that:
- online is 1
- type is USB

Change-Id: I6fd7328477977d30aa4ff66d5ffbd923067e3117
Signed-off-by: Sameer Nanda <snanda@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/276877
Reviewed-by: Alec Berg <alecaberg@chromium.org>