CHROMIUM: drm: rockchip/dw_hdmi-rockchip: Call drm_encoder_cleanup() in error path
The drm_encoder_cleanup() was missing both from the error path of
dw_hdmi_rockchip_bind().  This caused a crash when slub_debug was
enabled and we ended up deferring probe of HDMI at boot.

This call isn't needed from unbind() because if dw_hdmi_bind() returns
no error then it takes over the job of freeing the encoder (in
dw_hdmi_unbind).

BUG=chrome-os-partner:41376
TEST=w/ series, boot mickey with slub_debug=FZPUA and see no crash.

Change-Id: I65b254768514a4c0e80fdec07dd001dde948a1a4
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/277152
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>