CHROMIUM: drm/rockchip: Allow slop in clk_set_rate
If clk_set_rate gave us a slightly less rate than the one we requested, we try
to see if there's anything closer above that rate, by requesting rate + .01%.
If we find something closer that way, we'll automatically use it, otherwise
clk_set_rate will fallback to the original rate it returned.

The upstream patch for this would have to be implemented differently since
it's cleaner to use clk_set_rate_range there.

BUG=chrome-os-partner:41190
TEST=See test for "redo rockchip hdmi to allow slop"

Change-Id: Ie379275f6fbb15434ce8d4bef660d32668c5a586
Signed-off-by: Doug Anderson <dianders@chromium.org>
Signed-off-by: Alexandru M Stan <amstan@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/281004