FROMLIST: drm: bridge/dw_hdmi: Filter modes > 165MHz for DVI
If you plug in a DVI monitor to your HDMI port, you need to filter out
clocks > 165MHz.  That's because 165MHz is the maximum clock rate that
we can run single-link DVI at.

If you want to run high resolutions to DVI, you'd need some type of an
active adapter that pretended that it was HDMI, interpreted the
signal, and produced a new dual link DVI signal at a lower clock rate.

BUG=chrome-os-partner:41189
TEST=Plug into DVI monitor and see modes get filtered

(am from https://patchwork.kernel.org/patch/6630131/)
Change-Id: Id2ce7d1ac23f431d455ffcadfdae2b84cad3655a
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/278255
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>