UPSTREAM: clk: rockchip: protect register macros against multipart values
The macros calculate the correct offset from the id in the parameter.
If this parameter does not consist of a single number, the calculation
will return wrong registers in the best case or create unaligned accesses
in the worst case. So protect the calculations against such values.

BUG=chrome-os-partner:42869
TEST=Build and boot

Change-Id: I5db5c27805c03b08bbc15aef083c9dc7256b462b
Signed-off-by: Heiko Stuebner <heiko@sntech.de>
Signed-off-by: Stephen Boyd <sboyd@codeaurora.org>
Signed-off-by: Doug Anderson <dianders@chromium.org>
(cherry picked from git.kernel.org clk/linux.git clk-next
 commit 2d7884a7d0e70f9509881c40eaee3f5a5c201b07)
Reviewed-on: https://chromium-review.googlesource.com/287292
Reviewed-by: Alexandru Stan <amstan@chromium.org>