Revert "CHROMIUM: usb: introduce way to determine if wakeup is enabled for a port"
This reverts commit 33a9945543ba27d2c87dd777000dd7f1764d8198.

Nobody is using this function in the chromeos-3.14 tree.  The function
was intended to be sent upstream and used by the tegra xhci driver but
it sounds like the new consensus is to export
wakeup_enabled_descendants() and use that.

Revert the CL so we don't accidentally use it and also to give us a
clean pick for the CL that exports wakeup_enabled_descendants().

BUG=chrome-os-partner:40114
TEST='git grep usb_port_may_wakeup' => nothing; Build and boot

Change-Id: I1814b1bd621f58b1def92f7bcb9ec6507b87b5f3
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/281041
Reviewed-by: Andrew Bresticker <abrestic@chromium.org>