CHROMIUM: drm: rockchip/dw_hdmi-rockchip: add slop to more tables
In the change (bbf6323 CHROMIUM: drm: rockchip/dw_hdmi-rockchip: redo
rockchip hdmi to allow slop) we forgot to add clock slop to some of
the tables.  That means that if we pick a clock that is just slightly
higher than one of the ones we expected, we'll get the wrong values.
Fix it.

This was found by code inspection.

BUG=chrome-os-partner:41190
TEST=Try a bunch of HDMI modes

Change-Id: Ib64cd685170cd6ad11130fb64c361be937a97dfb
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/281823
Reviewed-by: Alexandru Stan <amstan@chromium.org>