CHROMIUM: usb: dwc2: Add TODO comments in dwc2_port_resume()
We really ought to have some spinlocks in there.  ...but there's a
sleep.  ...but the sleep appears to be longer than is really needed.
I'm not going to solve this all now, so TODOs for now.  Hopefully
someone from Rockchip can carry this upstream and figure out what to
do.

BUG=chrome-os-partner:41748
TEST=w/ series USB is happy, even with s2r

Change-Id: Iaa73fc2ce72f8da6364187dca51ee8e1164de9a8
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/282078