CHROMIUM: usb: dwc2: Make bus suspend/resume do something for dwc2
It's important to actually suspend any active ports when doing a bus
suspend/resume.  We should do that.  Apparently it's also important to
turn off the clock if there's nothing plugged in.

This patch is pretty much just a rehash of (0cf884e usb: dwc2: add bus
suspend/resume for dwc2), which was reverted upstream because it broke
things on some SoCs.  Someone really needs to resolve the problems and
get this landed upstream.

There are also several places that seem like they need spinlocks, but
for now I'm just not going to change things since I don't want to
break things further.  Again: this should be sorted out upstream.

BUG=chrome-os-partner:41748
TEST=w/ series USB is happy, even with s2r

Change-Id: I3dcb6a0403e73562c7d2a979df666ea91f3c6a75
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/282079