CHROMIUM: drm: rockchip/dw_hdmi-rockchip: Add clks: 25.2, 68.25, 72, 79.5
Add some more modes based on existing rates in the PLL table and based
on the PHY documentation.

New rates and anticipated PLL settings:
- 25.2 MHz                             =>  126 MHz / 5
- 68.25 MHz                            => 2184 MHz / 32
- 72 MHz                               =>  216 MHz / 3
- 79.5 MHz                             => 1272 MHz / 16

...if we want to run the PLL slower we may be able to introduce some
low jitter settings that are roughly the same as the ones above.  For
now I'm just using existing PLL rates.

Note that and 25.2 MHz is not in the PHY table, but 25.175 and 27 are
so I figured out rates based on that.

BUG=chrome-os-partner:41190
TEST=Test with HDMI analyzer; plug into monitors

Change-Id: I9774ae2c7814be353a0a390c62e47f5c861bbac5
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/284162