CHROMIUM: drm: rockchip: edp: Handle states other than just "ON" and "OFF" better
The rockchip_edp_poweroff() function was optimized to be a noop if the
state was already "OFF".  ...but we use the rockchip_edp_poweroff()
for more than just off: we also use it for "STANDBY" and "SUSPEND".

Really the function ought to check for "not ON".

BUG=chrome-os-partner:42051
TEST=See bug

Change-Id: Iaa85fa38083deb0c16643bdd1b1be34b169c94e4
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/283566
Reviewed-by: 智情 姚 <mark.yao@rock-chips.com>
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>
Commit-Queue: Daniel Kurtz <djkurtz@chromium.org>
Trybot-Ready: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>