CHROMIUM: clk: rockchip: Revert more questionable PLL rates
The PLL rates added in the commit (5a380cc CHROMIUM: clk: rockchip:
expand more clocks support) don't appear to have been tested properly
for jitter.

We already reverted one of the rates in (21ffb85 CHROMIUM: clk:
rockchip: take 2.130 GHz clock out).  Let's get rid of the rest.

Looking at the goals of the original CL and places where we handled
those rates:

    PLL Rate    --->    TMDS

    1758  MHz           73.25  MHz
    78.75 MHz           78.75  MHz
      => Less jittery rates now:
         CL:Ie391e6babbb8496c0cd61b419a4d62e213859684

    115.5 MHz           115.50 MHz
    119.0 MHz           119.00 MHz
    972.0 MHz           162.00 MHz
      => Removed until we can solve jitter:
         http://crosreview.com/284163

    2130  MHz           88.75  MHz
      => Less jittery rate now:
         http://crosreview.com/282355

Note that the orignal change also removed the rate (48000000, 1, 64,
32) from the table.  I have no idea why that was squashed in there,
but that rate was invalid anyway (it appears to have an out of bounds
NO).  I'm not putting that rate in.  Thus this plus the removal of the
2.130 GHz rate are NOT a complete revert of the original change).

BUG=chrome-os-partner:41882
TEST=Measure jtter many times

Change-Id: I40c84d38e03f4c38dedefee94cafe6e426e76fba
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/284595
Reviewed-by: Alexandru Stan <amstan@chromium.org>