CHROMIUM: clk: rockchip: Use a better jitter clock for 300MHz PLL
The 300MHz PLL rate gets chosen for NPLL when we want 75MHz HDMI
clocks.  The old rate was a bit jittery.  This new rate has a higher
(but still in range) Fvco, which reduces jitter.

After this change the HDMI TV in the lab can hold onto a 75MHz HDMI
signal OK.

BUG=chrome-os-partner:41190
TEST=Check for jitter in lab; try on real monitor and see it's fixed.

Change-Id: Ia44d67d42ec32ed5992e40d6862f423b3854e2f3
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/284574
Reviewed-by: Alexandru Stan <amstan@chromium.org>