CHROMIUM: drm: rockchip/dw_hdmi-rockchip: Better 85.5 and 88.75
Adjust the rates we pick for 85.5 and 88.75 to be better for jitter.

We already adjusted 88.75 once to choose 88.8, but it turns out that
there's an even better rate to use: 88.695653.  This is slightly
further away from the target rate (though not by much) but since it
has a higher Fvco we get better overall jitter.  This is still
uncomfortably close to the margin, but the improvment here appears to
be enough to fix one monitor I tested.

We never adjusted 85.5 before.  The old rate was out of spec for
jitter.  The new one is neatly in spec.  This change also fixes a real
monitor.

BUG=chrome-os-partner:41190
TEST=Check for jitter in lab; try on real monitor and see it's fixed.

Change-Id: Ic639f77223733785921381a6c9f896314615fd12
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/284575
Reviewed-by: Alexandru Stan <amstan@chromium.org>