CHROMIMU: ARM: dts: add mickey thermal strategy
According test, cpu frequency 1.2G, GPU 400M will enough for play
1080p@60fps online video, so we will try very hard to keep at least
that frequency.

As part of our strategy, we take into account that the GPU will heat
the CPU and the CPU will heat the GPU.  Thus we'll use the CPU as a
cooling device for the GPU and vice versa.  Note that we generally try
to throttle the GPU slightly earlier than the CPU, which in testing
seems to give better results.

BUG=chrome-os-partner:41851
TEST=play 1080p@60fps online video, the thermal work we expect

Change-Id: I11b447702e2627e848dd572c138d833ded11d94e
Signed-off-by: huang lin <hl@rock-chips.com>
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/284569