CHROMIUM: drm: rockchip/dw_hdmi-rockchip: Remove 66MHz
The 66MHz clock rate has been in the HDMI driver since its first
version, but its PHY settings don't match anything else and don't seem
to make any sense.  These settings are also not present in the
databook for the PHY.  I have been unable to locate any monitors that
use this clock rate.

Let's remove the settings for now.  If we can find a way to test this
clock rate properly we can add it back in, presumably with correct
settings.

BUG=chrome-os-partner:42595
TEST=HDMI rates near 66MHz still work fine.

Change-Id: I1365ebe760b6caa59f5f8784e1efb2e4b1f08fdc
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/285243