CHROMIUM: drm: rockchip/dw_hdmi-rockchip: Improve jitter
We improve jitter for rates > 69 MHz (other than 74.24 and 148.5).
That means we improve jitter for:

* 71 MHz
* 72 MHz
* 73.25 MHz
* 74.44 MHz
* 75 MHz
* 78.75 MHz
* 79.5 MHz
* 83.5 MHz
* 85.5 MHz
* 88.75 MHz

...some of these rates were already OK and this just improves the
margin.  A few of them were borderline and this may improve behavior
on monitors.

Jitter was improved by lowering the MPLL bandwidth to account for high
frequency noise in the rk3288 PLL.  In each case MPLL bandwidth was
lowered only enough to get us a comfortable margin.  We believe that
lowering the bandwidth like this is safe given sufficient testing.

Note that this adjustment comes with a matching adjustment to the
python that generates these tables.  See http://crosreview.com/286941

BUG=chrome-os-partner:41190
TEST=Test jitter a lot; test on a lot of monitors.

Change-Id: Ia429e12e02ce6af4c91bd243cce36490fa056b85
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/286950
Reviewed-by: Alexandru Stan <amstan@chromium.org>
Commit-Queue: Yakir Yang <ykk@rock-chips.com>
Tested-by: Yakir Yang <ykk@rock-chips.com>