CHROMIUM: drm: rockchip/dw_hdmi-rockchip: Add 119.0 and 136.75
Add 119.0 MHz and 136.75 MHz.  These rates require us to increase our
slop margin a little bit, but not enough to really matter too much.
For instance: the period is increased by 23.6 ps due to the error
between 119 MHz and 118.667 MHz.  The measured jitter is ~120 ps.
Thus we still end up with a total jitter of 145 ps which is easily
within margin.

BUG=chrome-os-partner:41190
TEST=Test clock for jitter on scope; try on real monitors / TVs

Change-Id: I9b36c3b1d96e88b229e3e77a01127fe0041ce203
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/287296