CHROMIUM: drm: rockchip/dw_hdmi-rockchip: Add rates >= 101 MHz
Now that we have a solution for jitter, add back in some useful modes
at 101 MHz and above.  Specifically:
* 101.0  MHz
* 106.5  MHz
* 108.0  MHz
* 115.5  MHz
* 135.0  MHz
* 297.0  MHz

Note that rates above 135.0 MHz (other than the special 297 MHz) were
not added because we lose control of jitter again.  If we want to add
more rates we need to increase the Fvco limit of the rk3288 PLL and do
optimization (like was done for 594 MHz).

BUG=chrome-os-partner:41190
TEST=Test jitter a lot; test on a lot of monitors.

Change-Id: I8a28517402605d8812b4d955800db907b6bce2ad
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/287295