HACK: drm: rockchip/dw_hdmi-rockchip: Hack out large resolutions
On rk3288:
* VOP BIG max output resolution: 3840x2160
* VOP LITTLE max output resolution: 2560x1600

Ideally DRM would be able to query the VOP that was assigned and
filter out modes that are too big.  DRM has the ability to filter
modes at the connector level (mode_valid in HDMI) and at the FB level
(dev->mode_config.max_width and dev->mode_config.max_height), but
doesn't appear to have the ability to filter modes at the CRTC level
(which is what the VOP is).

The above is made worse by the fact that the HDMI connector isn't
guaranteed to be connected to a CRTC when mode_valid() is called.
...and after a CRTC is attached we're not guaranteed to call
mode_valid() again.  If either of those restrictions wasn't present
then we could co-opt the mode_valid() to add a call into the CRTC.

Until we can find a way to filter modes at the VOP level, we'll just
make the assumption that HDMI is using "VOP BIG" and limit the max
resolution there.

NOTE: right now we only might use VOP LITTLE for HDMI is if we're doing
hardware mirroring of the internal and external displays.  For that
case we can be certain that we're not over the limit so we're OK.  In
the future we could imagine a world where we'd dynamically use the
LITTLE/BIG VOP (if both are free) for HDMI depending on the actual
resolution being used at the moment.  That would allow us to save power
when we could but still support all resolutions.  None of that is
implmeneted at the moment, though, so this HACK should suffice.

BUG=chrome-os-partner:41190
TEST=High resoultion modes no longer show up.

Change-Id: Iab6174bf2ac02f3a1d5d866abf6c3b4546469847
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/287294
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>