CHROMIUM: cros_usbpd-charger: report dual-role as type USB when sourcing power
When we are a power source to another dual role device, the
firmware reports charger type as NONE.  The driver should report the
charger type as USB for this case so that the UI can show the
charger in the dropdown list.

BUG=chromium:513051
TEST=connect two dual-role devices together and verify that the the
power supply type is USB on both devices.

Change-Id: Ia559cdd0c8c97542a6c44fa4d53123b13d899a67
Signed-off-by: Sameer Nanda <snanda@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/287661
Reviewed-by: Alec Berg <alecaberg@chromium.org>