CHROMIUM: ARM: dts: 158 default for emmc (not 90) on veyron
The default phase is used both during the identification (400 kHz)
phase and also as the default phase if all phases pass during tuning.

Using 158 for ID phase is fine to use for eMMC because we actually
end up at ~90 anyway.  At 400kHz the delay elements really have no
effect because the max is 255 and 60ps * 255 hardly changes the phase
at all when we have 400kHz.

It turns out that on minnie that 90 degrees is a very bad phase to end
up with if all phases pass because even though 90 degrees will
sometimes pass tuning it's not a good enough phase to use for
everything.  In general tuning relies on detecting some phases that
fail so that we can get as far away as possible from them.  158 works
better as a default.

Note that 158 was the phase we were using for eMMC anyway on a number
of veyron devices that I spot checked.

BUG=chrome-os-partner:43068
TEST=Reboot tests; suspend/resume tests

Change-Id: Ia6bb52d70414b0129671b5a4b90f39e16c40cd3a
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/288431
Reviewed-by: Caesar Wang <wxt@rock-chips.com>
Reviewed-by: Alexandru Stan <amstan@chromium.org>
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>