CHROMIUM: clk: rockchip: Allow more precision for some mmc clock phases
Because of the inexact nature of the extra MMC delay elements (it's
not possible to keep the phase monotonic and to also make phases (mod
90) > 67.5), we previously only allowed phases (mod 90) of 22.5, 45,
and 67.5.

There's actually no reason to limit it quite so much.  We can easily
make many more phases < 67.5 and still keep things monotonic.  Let's
do that.

NOTE: it's possible that later we may want to relax the monotonic
requirement.  Previously that seemed like a good idea, but in the case
of MMC really the point is to find bad phases and get as far asway
from the as possible.  If we get to test extra phases by going
slightly non-monotonic then that might be fine.  Worst case we would
end up at a phases that's slight differnt than the one we wanted, but
at least we'd still be quite far away from the a bad phase.  Anyway,
that's for a future change.

BUG=chrome-os-partner:43068
TEST=Enable dbg print in this file + see phases are good; reboot tests

Change-Id: I06bdb9f37093a48d94bee264a588b497dcb9fcf9
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/288580
Reviewed-by: Alexandru Stan <amstan@chromium.org>
Tested-by: Caesar Wang <wxt@rock-chips.com>
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>