CHROMIUM: ARM: dts: Add memory controller throttling for DDR3 veyriants
We'll try to use the memory controller to cool down veyron devices.

Right now we don't throttle too aggressively.  We'll go down to 533
when we get a little hot on DDR3 devices.

We probably should also go down to 333 when we get even hotter, but at
the moment that will cause problems because we're unlikely to be able
to run at 1.2 GHz (the minimum needed to change DDR freq) when we're
that hot.  We might want to re-tune the overall throttling strategy to
be more like rk3288-mickey eventually.

BUG=chrome-os-partner:42854
TEST=Run webgl aquarium and run watchspeed script

Change-Id: If5c4112f4958947020fd5bace0f3a7112a988b92
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/289597