CHROMIUM: ARM: dts: Adjust mickey thermals, including DMC throttle
This CL tweaks the mickey thermals, including:
* Throttle DMC to 433 on mickey unless we're pretty cool.
* We'll try not to throttle down the CPU below 1.4GHz unless we're
  pretty hot.
* We'll try not to throttle down the GPU below 400MHz unless we're
  pretty hot.
* Once we're pretty hot, try to gracefully degrade a bit.  We'll bring
  memory controller down to at most 333MHz (mostly because it's hard
  to slow it down more than that after CPU is running slow).
* We'll try not to throttle down to the "extra low" CPU speeds (where
  we don't save voltage) unless we're really desperate.

This CL could possibly result in mickey running just a tad hotter than
it used to since we don't throttle quite as aggressively until 82C (we
used to at 80C), but if those super agressive throttles were really
needed then something was probably pretty wrong anyway (like maybe
someone's aiming a hair dryer at the device).

Note that with this CL it's possible that we might get slower
performance just a little sooner than we used to also since we now
start throttling the DMC to 333 at 78 (and DMC 333 may not be enough
for great performance).  If approved, we'll raise this limit a little
in CL:I926ed5b86cecea25859be95740ba82500c5059ab

BUG=chrome-os-partner:42854
TEST=Run webgl aquarium and run watchspeed script

Change-Id: I926ed5b86cecea25859be95740ba82500c5059ab
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/289598
Reviewed-by: Derek Basehore <dbasehore@chromium.org>