UPSTREAM: ARM: rockchip: set correct stabilization thresholds in suspend
Currently the stabilization thresholds for the oscillator and external pmu
are statically set to 30ms based on a 32kHz clock rate. This leaves out the
case when we don't switch to the 32kHz clock when only entering the shallow
suspend mode where the logic keeps running.

So, set the correct threshold after we have determined if we switch to the
32kHz clock or stay with the 24MHz one. Also set the oscillator-
stabilization to 0 if it is kept running during suspend, as it of course
does not need to stabilize then.

Reported-by: Chris Zhong <zyw@rock-chips.com>
Signed-off-by: Heiko Stuebner <heiko@sntech.de>
Reviewed-by: Chris Zhong <zyw@rock-chips.com>
Tested-by: Chris Zhong <zyw@rock-chips.com>
Reviewed-by: Douglas Anderson <dianders@chromium.org>
(cherry-picked from git-kernel.org mmind/linux-rockchip.git v4.3-armsoc/soc
commit d1d3a1a1d745095261d4a927c0c06588e5bc7f6f)

BUG=chrome-os-partner:43047
TEST=echo 0 > /sys/module/pm/parameters/deep_sleep
powerd_dbus_suspend
press keyboard can wakeup

Change-Id: I797acb819c6199c832f5e6f74c9be492db1d4d92
Signed-off-by: Chris Zhong <zyw@rock-chips.com>
Reviewed-on: https://chromium-review.googlesource.com/292621
Reviewed-by: Douglas Anderson <dianders@chromium.org>