CHROMIUM: ARM: dts: rialto: Add modem power pinctrl
The modem's shutdown pin is hiZ during booting until some userspace tool resets
the modem. This causes weird behavoir at bootup. Sometimes the modem enumerates,
sometimes it doesn't, sometimes it confuses the ehci host and it'll never
enumerate until the ehci driver gets restarted.

We ought to just drive those lines properly (preferrably to keep the modem off)
until we actually need the modem.

Turns out that in the common veyron.dtsi file we use gpio 4 26 for bt_dev_wake
(and it assumes all inhereting boards will have it too), this is not the
case for rialto, so it needs to have the bt_dev_wake_* nodes deleted.

BUG=chrome-os-partner:43271
TEST=Note how the modem doesn't enumerate at bootup anymore
(~2 seconds after boot)
It enumerates only when the lines are set so it turns on, consistently!

Change-Id: If1d21ec2705e4ca6cb76ab42920bc96c930075b0
Signed-off-by: Alexandru M Stan <amstan@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/290631
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Karl Townsend <karlt@chromium.org>