FROMLIST: drm: bridge/dw_hdmi: replace CTS calculation for the ACR
Given the TDMS clock, audio sample rate, and the N parameter, we can
calculate the CTS value for the audio clock regenerator (ACR) using the
following calculation given in the HDMI specification:

	CTS = ftdms * N / (128 * fs)

The specification says that the CTS value is an average value, which is
true if the source hardware measures it.  Where source hardware needs it
to be programmed, it is particularly difficult to alternate it between
two values correctly to ensure that we achieve a correct "average"
fractional value at the sink.

Also, there's the problem that our "ftdms" is not a fully accurate
value; it is rounded to a kHz value.  This introduces an unnecessary
(and harmless) fractional value into the above equation for combinations
like 148.5MHz/1.001 for 44100Hz - we still calculate the correct CTS
value.

Conflicts:
    drivers/gpu/drm/bridge/dw_hdmi.c
We appeared to have an earlier version of patches.

BUG=chrome-os-partner:44695
TEST=Audio works w/ lots of resolutions

Change-Id: I10573fb55d21ba5075d60522adc7199377e3b128
Signed-off-by: Russell King <rmk+kernel@arm.linux.org.uk>
Signed-off-by: Douglas Anderson <dianders@chromium.org>
(am from https://patchwork.kernel.org/patch/6975261/)
Reviewed-on: https://chromium-review.googlesource.com/297618
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
(cherry picked from commit 61769485c5a6ada407165263a409ba24668f5e17)
Reviewed-on: https://chromium-review.googlesource.com/299468