CHROMIUM: ARM: rockchip: Hack in a rk3288 panic notifier to sample PC
This implements the old "print the DBGPCSR at panic time" on the
rk3288.  This is pretty much just a copy-paste job from
<https://chromium-review.googlesource.com/#/c/66196/> but changed for
the rk3288 address.

This is obviously terribly ugly code.  We'll have to decide whether to
land this or try to find something that will make upstream happy.
Note that there is very little rk3288-specific here except the clocks
and the base address.  The debug register we're looking at is generic
to all A17s and the whole thing worked the same on A15 too.  That all
means that there is likely to be much bikeshedding about how exactly
to do this.  ...and maybe this needs to fit into a larger module for
accessing the debug registers?  Certainly upstream has a "coresight"
device tree property.

BUG=chrome-os-partner:42868
TEST=echo HARDLOCKUP > /sys/kernel/debug/provoke-crash/DIRECT

Change-Id: I3b163f61286df752d3a14003ff190170a58aaa7e
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/292319
Reviewed-by: Andrew Bresticker <abrestic@chromium.org>