CHROMIUM: clk: rockchip: Register some debug clocks as global
In (CHROMIUM: ARM: rockchip: Hack in a rk3288 panic notifier to sample
PC) we'd like to access a few clocks without any associated device.
That's pretty ugly and gross, but since upstream is likely to have a
very different solution it doesn't seem to make sense to try to
architect something cleaner just for our tree.

We'll instead just register a few global clocks.  If this were common
we would move it to a common rockchip clock file, but since this is
just for something local to our tree we'll put it in the rk3288 file.

BUG=chrome-os-partner:42868
TEST=Used by CL:I3b163f61286df752d3a14003ff190170a58aaa7e

Change-Id: Ie83bdf976cdb39fb2123af58f42c9ceac957d320
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/292860
Reviewed-by: Andrew Bresticker <abrestic@chromium.org>