CHROMIUM: ARM: rockchip: Try to read PC a few times in panic notifier
Since we can't get a real stack crawl if we can't stop the other CPU,
let's at least sample the PC a few times in case that gives us a
better clue of what's going on.

We'll also skip printout out the current PC since we know exactly
where we'll be.

BUG=chrome-os-partner:42868
TEST=Modify HARDLOCKUP test to do a few things and see multiple PCs

Change-Id: I6f9c62415f70fb6ed60cf0cdb18924c8c561e9c2
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/292861
Reviewed-by: Andrew Bresticker <abrestic@chromium.org>