CHROMIUM: clk: rockchip: Add highish Fvco 273.6 MHz and 238 MHz
We've decided that it makes sense to go up above 2.2GHz for Fvco for
some clock rates for HDMI jitter purposes.  Now that we've decided
that, let's go ahead and use this to improve our ability to make 119
MHz and 136.75 MHz pixel clocks.  Previously we used some rates that
were a bit far off from the rates that we wanted.  Now we can get much
closer rates that are still good jitter (actually, they're slightly
better).

Note that we still can't actually make 136.75 with 273.6, but we can
make 136.8 which is much closer than the 136.5 we were making before.

BUG=chrome-os-partner:43781
TEST=Measure jitter; try on real monitor/TV

Change-Id: I9b98bac6bdbe34d38a069c22e3f0b8efadc317cd
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/295303
(cherry picked from commit 418c565e79536100203c15a0f81961414643b899)
Reviewed-on: https://chromium-review.googlesource.com/296031