CHROMIUM: drm: rockchip/dw_hdmi-rockchip: Set cur_ctr to 0 always
We're already setting cur_ctr to 0 for higher clock rates, and really
this setting seems to affect different types of jitter.

Let's do a bunch of testing with this at 0 everywhere and if jitter
looks good, let's take it.

I've found at least one case where this helps.  When using an HDMI to
VGA adapter to a LCD monitor, the signal was bad before this change
and looks good after this change.

BUG=chrome-os-partner:44569
TEST=Try lots of rates across lots of monitors.

Change-Id: If2e6862cf0cfaa5085feb04f6d5b63728371157d
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/295708
Tested-by: Yakir Yang <ykk@rock-chips.com>
Reviewed-by: Yakir Yang <ykk@rock-chips.com>
(cherry picked from commit 516b4fa23d914049758722bafa7703165b167fd7)
Reviewed-on: https://chromium-review.googlesource.com/299469