CHROMIUM: drm: bridge/dw_hdmi: Get rid of special cases for N
Most of the special cases for N in the HDMI spec assume we're making a
very very specific rate (down the the Hz).  We're not.  Get rid of those
special cases.

Note in a future change we'll add back in a few special cases for
297MHz+.

BUG=chrome-os-partner:44695
TEST=Audio works w/ lots of resolutions

Change-Id: Iea7e7b372e6c66751487b36dbcda008666864644
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/297619
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
(cherry picked from commit 1bc863a891f8e69e3747ab78636045218d2525d2)
Reviewed-on: https://chromium-review.googlesource.com/299751